+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/events/AbstractStrategy.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.strategy;
+package uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.events;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
