+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/events/Strategy.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.strategy;
+package uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.events;
 
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.model.Envelope;
